backend/requirements.txt
Django==6.0.2
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
django-cors-headers==4.9.0
django-environ==0.13.0
django-redis==5.4.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
pillow==12.1.1
